updated hrp helpers, set DefaultAddressHrp on sdk-js LibraryConfig
